Fix draft_orders: drop failed table, recreate with INTEGER FKs matching medications table
Previous attempts left a broken table. Migration now drops and recreates.
Fallback CREATE TABLE in endpoints uses no FK constraints to avoid type issues.

+++ netlify/functions/draft-orders-get.js
@@ -10,11 +10,12 @@ exports.handler = async (event) => {
     const status = params.status || 'pending_review';
 
     // Ensure table exists
+    // Ensure draft_orders table exists (idempotent, no FKs to avoid type issues)
     await db.query(`
       CREATE TABLE IF NOT EXISTS draft_orders (
         id SERIAL PRIMARY KEY,
-        medication_id BIGINT NOT NULL,
-        location_id BIGINT,
+        medication_id INTEGER NOT NULL,
+        location_id INTEGER,
         current_stock_boxes NUMERIC(10,2) NOT NULL DEFAULT 0,
         min_level_boxes INTEGER NOT NULL DEFAULT 0,
         suggested_quantity INTEGER NOT NULL,
@@ -23,8 +24,8 @@ exports.handler = async (event) => {
         intelligence_snapshot JSONB,
         source VARCHAR(20) NOT NULL DEFAULT 'auto',
         status VARCHAR(20) NOT NULL DEFAULT 'pending_review',
-        generated_by BIGINT,
-        approved_by BIGINT,
+        generated_by INTEGER,
+        approved_by INTEGER,
         generated_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT NOW(),
         approved_at TIMESTAMP WITH TIME ZONE,
         rejected_at TIMESTAMP WITH TIME ZONE,
